refactor(deeplink): update package names and imports
This commit refactors the deep link module by updating the package names and imports to use `co.anitrend.android.deeplink` instead of `co.anitrend.deeplink`.

The changes include:

-   Updating package names in all source files to `co.anitrend.android.deeplink`.
-   Updating import statements across all files to reflect the new package structure.
-   Updating references to classes and modules within the codebase to match the new package names.

These changes improve the organization and consistency of the deep link module's codebase.
