+++ android/deeplink/src/main/kotlin/co/anitrend/android/deeplink/component/presenter/SplashPresenter.kt
@@ -14,7 +14,7 @@
  *     You should have received a copy of the GNU General Public License
  *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package co.anitrend.deeplink.component.presenter
+package co.anitrend.android.deeplink.component.presenter
 
 import android.content.Context
 import co.anitrend.android.core.settings.Settings
