+++ android/deeplink/src/main/kotlin/co/anitrend/android/deeplink/initializer/FeatureInitializer.kt
@@ -14,12 +14,12 @@
  *     You should have received a copy of the GNU General Public License
  *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package co.anitrend.deeplink.initializer
+package co.anitrend.android.deeplink.initializer
 
 import android.content.Context
 import co.anitrend.core.initializer.contract.AbstractFeatureInitializer
 import co.anitrend.core.koin.helper.DynamicFeatureModuleHelper.Companion.loadModules
-import co.anitrend.deeplink.koin.moduleHelper
+import co.anitrend.android.deeplink.koin.moduleHelper
 
 class FeatureInitializer : AbstractFeatureInitializer<Unit>() {
     /**
