+++ android/deeplink/src/main/kotlin/co/anitrend/android/deeplink/koin/Modules.kt
@@ -14,48 +14,48 @@
  *     You should have received a copy of the GNU General Public License
  *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package co.anitrend.deeplink.koin
+package co.anitrend.android.deeplink.koin
 
 import android.content.Intent
 import co.anitrend.android.core.environment.IAniTrendEnvironment
 import co.anitrend.core.koin.helper.DynamicFeatureModuleHelper
 import co.anitrend.data.user.settings.IUserSettings
-import co.anitrend.deeplink.component.presenter.OnBoardingPresenter
-import co.anitrend.deeplink.component.presenter.SplashPresenter
-import co.anitrend.deeplink.component.route.AboutRoute
-import co.anitrend.deeplink.component.route.ActivityRoute
-import co.anitrend.deeplink.component.route.AiringRoute
-import co.anitrend.deeplink.component.route.CharacterRoute
-import co.anitrend.deeplink.component.route.DiscoverRoute
-import co.anitrend.deeplink.component.route.EpisodesRoute
-import co.anitrend.deeplink.component.route.FallbackAction
-import co.anitrend.deeplink.component.route.ForumDiscoverRoute
-import co.anitrend.deeplink.component.route.ForumRoute
-import co.anitrend.deeplink.component.route.MainRoute
-import co.anitrend.deeplink.component.route.MediaListRoute
-import co.anitrend.deeplink.component.route.MediaRoute
-import co.anitrend.deeplink.component.route.NewsRoute
-import co.anitrend.deeplink.component.route.NotificationRoute
-import co.anitrend.deeplink.component.route.OAuthRoute
-import co.anitrend.deeplink.component.route.ProfileRoute
-import co.anitrend.deeplink.component.route.RecommendationRoute
-import co.anitrend.deeplink.component.route.ReviewRoute
-import co.anitrend.deeplink.component.route.SearchPageRoute
-import co.anitrend.deeplink.component.route.SearchRoute
-import co.anitrend.deeplink.component.route.SettingsRoute
-import co.anitrend.deeplink.component.route.SocialRoute
-import co.anitrend.deeplink.component.route.StaffRoute
-import co.anitrend.deeplink.component.route.StudioRoute
-import co.anitrend.deeplink.component.route.SuggestionsRoute
-import co.anitrend.deeplink.component.route.UpdatesRoute
-import co.anitrend.deeplink.component.route.UserFavouritesRoute
-import co.anitrend.deeplink.component.route.UserReviewRoute
-import co.anitrend.deeplink.component.route.UserRoute
-import co.anitrend.deeplink.component.route.UserStatsRoute
-import co.anitrend.deeplink.component.screen.DeepLinkScreen
-import co.anitrend.deeplink.component.viewmodel.DeepLinkViewModel
-import co.anitrend.deeplink.environment.AniTrendEnvironment
-import co.anitrend.deeplink.provider.FeatureProvider
+import co.anitrend.android.deeplink.component.presenter.OnBoardingPresenter
+import co.anitrend.android.deeplink.component.presenter.SplashPresenter
+import co.anitrend.android.deeplink.component.route.AboutRoute
+import co.anitrend.android.deeplink.component.route.ActivityRoute
+import co.anitrend.android.deeplink.component.route.AiringRoute
+import co.anitrend.android.deeplink.component.route.CharacterRoute
+import co.anitrend.android.deeplink.component.route.DiscoverRoute
+import co.anitrend.android.deeplink.component.route.EpisodesRoute
+import co.anitrend.android.deeplink.component.route.FallbackAction
+import co.anitrend.android.deeplink.component.route.ForumDiscoverRoute
+import co.anitrend.android.deeplink.component.route.ForumRoute
+import co.anitrend.android.deeplink.component.route.MainRoute
+import co.anitrend.android.deeplink.component.route.MediaListRoute
+import co.anitrend.android.deeplink.component.route.MediaRoute
+import co.anitrend.android.deeplink.component.route.NewsRoute
+import co.anitrend.android.deeplink.component.route.NotificationRoute
+import co.anitrend.android.deeplink.component.route.OAuthRoute
+import co.anitrend.android.deeplink.component.route.ProfileRoute
+import co.anitrend.android.deeplink.component.route.RecommendationRoute
+import co.anitrend.android.deeplink.component.route.ReviewRoute
+import co.anitrend.android.deeplink.component.route.SearchPageRoute
+import co.anitrend.android.deeplink.component.route.SearchRoute
+import co.anitrend.android.deeplink.component.route.SettingsRoute
+import co.anitrend.android.deeplink.component.route.SocialRoute
+import co.anitrend.android.deeplink.component.route.StaffRoute
+import co.anitrend.android.deeplink.component.route.StudioRoute
+import co.anitrend.android.deeplink.component.route.SuggestionsRoute
+import co.anitrend.android.deeplink.component.route.UpdatesRoute
+import co.anitrend.android.deeplink.component.route.UserFavouritesRoute
+import co.anitrend.android.deeplink.component.route.UserReviewRoute
+import co.anitrend.android.deeplink.component.route.UserRoute
+import co.anitrend.android.deeplink.component.route.UserStatsRoute
+import co.anitrend.android.deeplink.component.screen.DeepLinkScreen
+import co.anitrend.android.deeplink.component.viewmodel.DeepLinkViewModel
+import co.anitrend.android.deeplink.environment.AniTrendEnvironment
+import co.anitrend.android.deeplink.provider.FeatureProvider
 import co.anitrend.navigation.DeepLinkRouter
 import com.kingsleyadio.deeplink.DeepLinkParser
 import com.kingsleyadio.deeplink.Environment
